+++ packages/core/src/components/HierarchyPicker/Hierarchy/HierarchyItem.tsx
@@ -94,7 +94,7 @@ class HierarchyItem extends React.Component<Props & WithStylesProps> {
       <>
         {selected && (
           <span {...css(styles.checkmark)}>
-            <IconCheckmark color={theme!.color.core.primary[3]} size={ICON_SIZE} decorative />
+            <IconCheckmark decorative color={theme!.color.core.primary[3]} size={ICON_SIZE} />
           </span>
         )}
 
@@ -117,7 +117,7 @@ class HierarchyItem extends React.Component<Props & WithStylesProps> {
         tabIndex={focused ? 1 : 0} // this is needed to find a focused parent item in a vertically aligned list
       >
         {this.renderItem()}
-        {item.items && <IconChevronRight size="1.4em" decorative inline />}
+        {item.items && <IconChevronRight decorative inline size="1.4em" />}
       </div>
     );
   }