+++ packages/core/src/components/IconButton.story.tsx
@@ -11,27 +11,27 @@ storiesOf('Core/IconButton', module)
   .add('A standard button, with different sizes.', () => (
     <>
       <IconButton onClick={action('onClick')}>
-        <IconCheck />
+        <IconCheck decorative />
       </IconButton>
       <br />
       <br />
       <IconButton onClick={action('onClick')}>
-        <IconCheck size="2em" />
+        <IconCheck decorative size="2em" />
       </IconButton>
     </>
   ))
   .add('An anchor link when passing `href`.', () => (
     <IconButton onClick={action('onClick')} href="https://github.com/airbnb/lunar" openInNewWindow>
-      <IconCheck />
+      <IconCheck decorative />
     </IconButton>
   ))
   .add('Wrapped in a tooltip.', () => (
     <IconButton onClick={action('onClick')} tooltip="This does something cool.">
-      <IconCheck />
+      <IconCheck decorative />
     </IconButton>
   ))
   .add('With a disabled state.', () => (
     <IconButton onClick={action('onClick')} disabled>
-      <IconCheck />
+      <IconCheck decorative />
     </IconButton>
   ));