+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.4.0 - 2019-06-07
+
+#### 🚀 Updates
+
+- **[withIcon]** Throw dev error if an ally prop is missing. (#84) ([75acae5](https://github.com/airbnb/lunar/commit/75acae5)), closes [#84](https://github.com/airbnb/lunar/issues/84)
+
+**Note:** Version bump only for package @airbnb/lunar-icons
+
+
+
+
+
 ## 1.3.0 - 2019-05-31
 
 #### 🚀 Updates